+++ server/src/test/java/org/opensearch/search/approximate/ApproximatePointRangeQueryTests.java
@@ -26,12 +26,13 @@ import org.apache.lucene.tests.index.RandomIndexWriter;
 import org.opensearch.search.internal.SearchContext;
 import org.opensearch.search.sort.SortOrder;
 import org.opensearch.test.OpenSearchTestCase;
+import org.opensearch.test.TestSearchContext;
 
 import java.io.IOException;
 
 import static java.util.Arrays.asList;
+import static org.opensearch.search.approximate.ApproximatePointRangeQuery.LONG_FORMAT;
 import static org.apache.lucene.document.LongPoint.pack;
-import static org.mockito.Mockito.mock;
 
 public class ApproximatePointRangeQueryTests extends OpenSearchTestCase {
 
@@ -59,11 +60,13 @@ public class ApproximatePointRangeQueryTests extends OpenSearchTestCase {
                     try {
                         long lower = RandomNumbers.randomLongBetween(random(), -100, 200);
                         long upper = lower + RandomNumbers.randomLongBetween(random(), 0, 100);
-                        Query approximateQuery = new ApproximatePointRangeQuery("point", pack(lower).bytes, pack(upper).bytes, dims) {
-                            protected String toString(int dimension, byte[] value) {
-                                return Long.toString(LongPoint.decodeDimension(value, 0));
-                            }
-                        };
+                        Query approximateQuery = new ApproximatePointRangeQuery(
+                            "point",
+                            pack(lower).bytes,
+                            pack(upper).bytes,
+                            dims,
+                            LONG_FORMAT
+                        );
                         Query query = LongPoint.newRangeQuery("point", lower, upper);
                         IndexSearcher searcher = new IndexSearcher(reader);
                         TopDocs topDocs = searcher.search(approximateQuery, 10);
@@ -99,11 +102,13 @@ public class ApproximatePointRangeQueryTests extends OpenSearchTestCase {
                     try {
                         long lower = 0;
                         long upper = 1000;
-                        Query approximateQuery = new ApproximatePointRangeQuery("point", pack(lower).bytes, pack(upper).bytes, dims) {
-                            protected String toString(int dimension, byte[] value) {
-                                return Long.toString(LongPoint.decodeDimension(value, 0));
-                            }
-                        };
+                        Query approximateQuery = new ApproximatePointRangeQuery(
+                            "point",
+                            pack(lower).bytes,
+                            pack(upper).bytes,
+                            dims,
+                            LONG_FORMAT
+                        );
                         IndexSearcher searcher = new IndexSearcher(reader);
                         TopDocs topDocs = searcher.search(approximateQuery, 10);
                         assertEquals(topDocs.totalHits, new TotalHits(1000, TotalHits.Relation.EQUAL_TO));
@@ -137,11 +142,15 @@ public class ApproximatePointRangeQueryTests extends OpenSearchTestCase {
                     try {
                         long lower = 0;
                         long upper = 45;
-                        Query approximateQuery = new ApproximatePointRangeQuery("point", pack(lower).bytes, pack(upper).bytes, dims, 10) {
-                            protected String toString(int dimension, byte[] value) {
-                                return Long.toString(LongPoint.decodeDimension(value, 0));
-                            }
-                        };
+                        Query approximateQuery = new ApproximatePointRangeQuery(
+                            "point",
+                            pack(lower).bytes,
+                            pack(upper).bytes,
+                            dims,
+                            10,
+                            null,
+                            LONG_FORMAT
+                        );
                         IndexSearcher searcher = new IndexSearcher(reader);
                         TopDocs topDocs = searcher.search(approximateQuery, 10);
                         assertEquals(topDocs.totalHits, new TotalHits(10, TotalHits.Relation.EQUAL_TO));
@@ -180,19 +189,16 @@ public class ApproximatePointRangeQueryTests extends OpenSearchTestCase {
                             pack(lower).bytes,
                             pack(upper).bytes,
                             dims,
-                            11_000
-                        ) {
-                            protected String toString(int dimension, byte[] value) {
-                                return Long.toString(LongPoint.decodeDimension(value, 0));
-                            }
-                        };
+                            11_000,
+                            null,
+                            LONG_FORMAT
+                        );
                         IndexSearcher searcher = new IndexSearcher(reader);
                         TopDocs topDocs = searcher.search(approximateQuery, 11000);
-                        assertEquals(topDocs.totalHits, new TotalHits(11000, TotalHits.Relation.EQUAL_TO));
+                        assertEquals(new TotalHits(11000, TotalHits.Relation.EQUAL_TO), topDocs.totalHits);
                     } catch (IOException e) {
                         throw new RuntimeException(e);
                     }
-
                 }
             }
         }
@@ -220,11 +226,15 @@ public class ApproximatePointRangeQueryTests extends OpenSearchTestCase {
                     try {
                         long lower = 0;
                         long upper = 100;
-                        Query approximateQuery = new ApproximatePointRangeQuery("point", pack(lower).bytes, pack(upper).bytes, dims, 10) {
-                            protected String toString(int dimension, byte[] value) {
-                                return Long.toString(LongPoint.decodeDimension(value, 0));
-                            }
-                        };
+                        Query approximateQuery = new ApproximatePointRangeQuery(
+                            "point",
+                            pack(lower).bytes,
+                            pack(upper).bytes,
+                            dims,
+                            10,
+                            null,
+                            LONG_FORMAT
+                        );
                         Query query = LongPoint.newRangeQuery("point", lower, upper);
                         ;
                         IndexSearcher searcher = new IndexSearcher(reader);
@@ -271,12 +281,9 @@ public class ApproximatePointRangeQueryTests extends OpenSearchTestCase {
                             pack(upper).bytes,
                             dims,
                             10,
-                            SortOrder.ASC
-                        ) {
-                            protected String toString(int dimension, byte[] value) {
-                                return Long.toString(LongPoint.decodeDimension(value, 0));
-                            }
-                        };
+                            SortOrder.ASC,
+                            LONG_FORMAT
+                        );
                         Query query = LongPoint.newRangeQuery("point", lower, upper);
                         ;
                         IndexSearcher searcher = new IndexSearcher(reader);
@@ -305,11 +312,7 @@ public class ApproximatePointRangeQueryTests extends OpenSearchTestCase {
     }
 
     public void testSize() {
-        ApproximatePointRangeQuery query = new ApproximatePointRangeQuery("point", pack(0).bytes, pack(20).bytes, 1) {
-            protected String toString(int dimension, byte[] value) {
-                return Long.toString(LongPoint.decodeDimension(value, 0));
-            }
-        };
+        ApproximatePointRangeQuery query = new ApproximatePointRangeQuery("point", pack(0).bytes, pack(20).bytes, 1, LONG_FORMAT);
         assertEquals(query.getSize(), 10_000);
 
         query.setSize(100);
@@ -318,11 +321,7 @@ public class ApproximatePointRangeQueryTests extends OpenSearchTestCase {
     }
 
     public void testSortOrder() {
-        ApproximatePointRangeQuery query = new ApproximatePointRangeQuery("point", pack(0).bytes, pack(20).bytes, 1) {
-            protected String toString(int dimension, byte[] value) {
-                return Long.toString(LongPoint.decodeDimension(value, 0));
-            }
-        };
+        ApproximatePointRangeQuery query = new ApproximatePointRangeQuery("point", pack(0).bytes, pack(20).bytes, 1, LONG_FORMAT);
         assertNull(query.getSortOrder());
 
         query.setSortOrder(SortOrder.ASC);
@@ -330,24 +329,18 @@ public class ApproximatePointRangeQueryTests extends OpenSearchTestCase {
     }
 
     public void testCanApproximate() {
-        ApproximatePointRangeQuery query = new ApproximatePointRangeQuery("point", pack(0).bytes, pack(20).bytes, 1) {
-            protected String toString(int dimension, byte[] value) {
-                return Long.toString(LongPoint.decodeDimension(value, 0));
-            }
-        };
+        ApproximatePointRangeQuery query = new ApproximatePointRangeQuery("point", pack(0).bytes, pack(20).bytes, 1, LONG_FORMAT);
 
         assertFalse(query.canApproximate(null));
 
-        ApproximatePointRangeQuery queryCanApproximate = new ApproximatePointRangeQuery("point", pack(0).bytes, pack(20).bytes, 1) {
-            protected String toString(int dimension, byte[] value) {
-                return Long.toString(LongPoint.decodeDimension(value, 0));
-            }
-
-            public boolean canApproximate(SearchContext context) {
-                return true;
-            }
-        };
-        SearchContext searchContext = mock(SearchContext.class);
+        ApproximatePointRangeQuery queryCanApproximate = new ApproximatePointRangeQuery(
+            "point",
+            pack(0).bytes,
+            pack(20).bytes,
+            1,
+            LONG_FORMAT
+        );
+        SearchContext searchContext = new TestSearchContext(null);
         assertTrue(queryCanApproximate.canApproximate(searchContext));
     }
 }